IN-134: As the access_logs argument is always configured but defaulted to (#22)
'false', there is no need to have it dynamic. Instead we configure it as
enabled depending on if the bucket name is configured. This will provide
clearer terraform plan output as we are seeing odd behaviour with the 'null'
alternative, plan's wanting to delete the access_logs attribute but not change it.

-       access_logs {
-           enabled = false -> null
        }