IN-327: PostgreSQL paramaeter group family naming contains only the (#26)
major version. When upgrading to 17.4 from 17, we found that using the version
as it is, created an invalid family name.

Therefore we are stripping any minor versions when specifying the family name in
the parameter group creation.

Acknowledged that this is very postgres specific and will not work for other engines
like mysql (where the families are further named e.g. mysql5.6). In this case it would
make more sense to pass in the family as a further parameter. This decision was toiled with,
but the metaflow module presumes that it will be deployed with postgres, so decided to
stick with also presuming postgres for this case.